(feat:calendar) changed the projects calculations from projects component to calendar component
