--- projects/resource-manager/src/lib/models/enums/availability.ts
@@ -1,3 +1,0 @@
-export enum Availability {
-  CUSTOM,
-}
