+++ projects/resource-manager/src/lib/projects/projects/projects.component.html
@@ -1,177 +1,13 @@
-<div ngClass="style" lib-projects>
+<!-- <div ngClass="style" lib-projects>
   <div [ngClass]="'calendar-info-container'">
     <p [ngClass]="'calendar-info-text'">Refine Search By</p>
     <p>
       <mat-icon> unfold_less</mat-icon>
     </p>
   </div>
-  <lib-calendar ngClass="calendar-container"></lib-calendar>
-</div>
-<div ngClass="filter-projects-container">
-  <lib-filter></lib-filter>
-  <div [ngClass]="'projects-container'" *ngIf="calendarView == 'month'">
-    <mat-accordion *ngFor="let campaign of currentMonthTasks">
-      <mat-expansion-panel
-        hideToggle
-        (opened)="campaignPanelOpenState = true"
-        (closed)="campaignPanelOpenState = false"
-        [ngClass]="'campaign-content'"
-        [ngStyle]="{
-          'margin-left': monthLeftSpace(campaign.startDate),
-          width: monthWidth(campaign.startDate, campaign.endDate),
-          position: 'relative',
-          'margin-bottom': '15px'
-        }"
-      >
-        <mat-expansion-panel-header>
-          <mat-panel-title
-            ><mat-icon>navigate_next</mat-icon> {{ campaign.campaignName }}
-          </mat-panel-title>
-          <mat-panel-description [ngClass]="'campaign-card'">
-            <div [ngClass]="'campaign-card'">
-              <p [ngStyle]="{ 'margin-right': '15px' }">
-                {{ campaign.startDate | date: "MMM d" }} -
-                {{ campaign.endDate | date: "MMM d" }}
-              </p>
-              <p>{{ campaign.name }}</p>
-            </div>
-            <p>{{ campaign.priority }} PRIORITY</p>
-          </mat-panel-description>
-        </mat-expansion-panel-header>
-        <div>
-          <mat-accordion *ngFor="let project of campaign.projectList">
-            <mat-expansion-panel
-              hideToggle
-              (opened)="projectPanelOpenState = true"
-              (closed)="projectPanelOpenState = false"
-              [ngStyle]="{ 'margin-bottom': '10px' }"
-            >
-              <mat-expansion-panel-header>
-                <mat-panel-title
-                  ><mat-icon>navigate_next</mat-icon>
-                  {{ project.projectName }}</mat-panel-title
-                >
-                <mat-panel-description [ngClass]="'campaign-card'">
-                  <div [ngClass]="'campaign-card'">
-                    <p [ngStyle]="{ 'margin-right': '15px' }">
-                      {{ project.startDate | date: "MMM d" }} -
-                      {{ project.endDate | date: "MMM d" }}
-                    </p>
-                    <p>{{ project.name }}</p>
-                  </div>
-                  <p>{{ project.priority }} PRIORITY</p>
-                </mat-panel-description>
-              </mat-expansion-panel-header>
-              <div [ngClass]="'task-container'">
-                <div *ngFor="let task of project.taskList">
-                  <mat-card
-                    [ngClass]="'task-cards'"
-                    (click)="openEmployeeDialog()"
-                  >
-                    <mat-card-content [ngClass]="'campaign-card'">
-                      <div>
-                        <p>{{ task.taskName }}</p>
-                        <p>10 Hrs</p>
-                        <p>
-                          {{ task.startDate | date: "MMM d" }} -
-                          {{ task.endDate | date: "MMM d" }}
-                        </p>
-                      </div>
-                      <div [ngClass]="'task-card-content'">
-                        <p>{{ task.priority }}</p>
-                        <p>{{ task.name }}</p>
-                      </div>
-                    </mat-card-content>
-                  </mat-card>
-                </div>
-              </div>
-            </mat-expansion-panel>
-          </mat-accordion>
-        </div>
-      </mat-expansion-panel>
-    </mat-accordion>
-  </div>
-  <div [ngClass]="'projects-container'" *ngIf="calendarView == 'week'">
-    <mat-accordion
-      *ngFor="let campaign of currentWeekTasks"
-      [ngClass]="'campaign-content'"
-    >
-      <mat-expansion-panel
-        hideToggle
-        (opened)="campaignPanelOpenState = true"
-        (closed)="campaignPanelOpenState = false"
-        [ngStyle]="{
-          'margin-left': weekLeftSpace(campaign.startDate),
-          width: weekWidth(campaign.startDate, campaign.endDate),
-          position: 'relative',
-          'margin-bottom': '10px'
-        }"
-      >
-        <mat-expansion-panel-header>
-          <mat-panel-title
-            ><mat-icon>navigate_next</mat-icon> {{ campaign.campaignName }}
-          </mat-panel-title>
-          <mat-panel-description [ngClass]="'campaign-card'">
-            <div [ngClass]="'campaign-card'">
-              <p [ngStyle]="{ 'margin-right': '15px' }">
-                {{ campaign.startDate | date: "MMM d" }} -
-                {{ campaign.endDate | date: "MMM d" }}
-              </p>
-              <p>{{ campaign.name }}</p>
-            </div>
-            <p>{{ campaign.priority }} PRIORITY</p>
-          </mat-panel-description>
-        </mat-expansion-panel-header>
-        <div>
-          <mat-accordion *ngFor="let project of campaign.projectList">
-            <mat-expansion-panel
-              hideToggle
-              (opened)="projectPanelOpenState = true"
-              (closed)="projectPanelOpenState = false"
-            >
-              <mat-expansion-panel-header>
-                <mat-panel-title
-                  ><mat-icon>navigate_next</mat-icon>
-                  {{ project.projectName }}</mat-panel-title
-                >
-                <mat-panel-description [ngClass]="'campaign-card'">
-                  <div [ngClass]="'campaign-card'">
-                    <p [ngStyle]="{ 'margin-right': '15px' }">
-                      {{ project.startDate | date: "MMM d" }} -
-                      {{ project.endDate | date: "MMM d" }}
-                    </p>
-                    <p>{{ project.name }}</p>
-                  </div>
-                  <p>{{ project.priority }} PRIORITY</p>
-                </mat-panel-description>
-              </mat-expansion-panel-header>
-              <div [ngClass]="'task-container'">
-                <div *ngFor="let task of project.taskList">
-                  <mat-card
-                    [ngClass]="'task-cards'"
-                    (click)="openEmployeeDialog()"
-                  >
-                    <mat-card-content [ngClass]="'campaign-card'">
-                      <div>
-                        <p>{{ task.taskName }}</p>
-                        <p>10 Hrs</p>
-                        <p>
-                          {{ task.startDate | date: "MMM d" }} -
-                          {{ task.endDate | date: "MMM d" }}
-                        </p>
-                      </div>
-                      <div [ngClass]="'task-card-content'">
-                        <p>{{ task.priority }}</p>
-                        <p>{{ task.name }}</p>
-                      </div>
-                    </mat-card-content>
-                  </mat-card>
-                </div>
-              </div>
-            </mat-expansion-panel>
-          </mat-accordion>
-        </div>
-      </mat-expansion-panel>
-    </mat-accordion>
-  </div>
-</div>
+  <lib-calendar
+    ngClass="calendar-container"
+    [projects]="projects"
+  ></lib-calendar>
+</div> -->
+<lib-calendar ngClass="calendar-container" [projects]="projects"></lib-calendar>
