+++ projects/resource-manager/src/lib/projects/projects/projects.component.scss
@@ -1,78 +1,3 @@
-.style {
-  display: flex;
-  flex-direction: row;
-}
-
-.calendar-info-text {
-  font-family: Roboto;
-  font-weight: 400;
-  color: #728192;
-}
-
-.calendar-info-container {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  color: #728192;
-  padding: 10px;
-  width: 13.5vw;
-  height: 44px;
-  border: solid 1px rgb(194, 193, 193);
-}
-
 .calendar-container {
   width: 85vw;
-}
-
-.filter-projects-container {
-  display: flex;
-  flex-direction: row;
-}
-
-.projects-container {
-  height: 605px;
-  overflow-y: auto;
-}
-
-// ----------------
-
-.mat-expanded .mat-expansion-panel-header-title .mat-icon {
-  transform: rotate(90deg);
-}
-
-.campaign-allocation-container {
-  padding-top: 90px;
-}
-
-.campaign-card {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
-
-.campaign-content {
-  font-size: 12px;
-  background-color: #eeeeee;
-}
-
-.task-card-content {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.task-container {
-  display: flex;
-  flex-direction: row;
-}
-
-.task-cards {
-  margin-right: 10px;
-  min-width: 220px;
-  max-height: 110px;
-  background-color: #ab94ff;
-  color: #ffffff;
-  font-family: Roboto;
-  cursor: pointer;
 }
