+++ projects/resource-manager/src/lib/services/filter/filter.service.ts
@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class FilterService {
-
-  constructor() { }
+  constructor() {}
 }
