+++ projects/resource-manager/src/lib/shared/employee-allocation/employee-allocation.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
-import { Availability } from '../../models/enums/availability';
 import { Resource } from '../../models/resources/resource';
 
 @Component({
@@ -27,26 +26,7 @@ export class EmployeeAllocationComponent implements OnInit {
   }
   startDate = new Date(1990, 0, 1);
   selectedValue: string = '';
-  resources: Resource[] = [
-    {
-      resourceId: 1,
-      resourceName: 'James',
-      resourceType: 'Chef',
-      availability: Availability.CUSTOM,
-    },
-    {
-      resourceId: 2,
-      resourceName: 'Andy',
-      resourceType: 'Driver',
-      availability: Availability.CUSTOM,
-    },
-    {
-      resourceId: 3,
-      resourceName: 'Sam',
-      resourceType: 'Chef',
-      availability: Availability.CUSTOM,
-    },
-  ];
+  resources: Resource[] = [];
   statuses: string[] = ['DEFINED', 'IN_PROGRESS'];
   priorities: string[] = ['LOW', 'MEDIUM', 'HIGH'];
   ngOnInit(): void {}
