+++ projects/resource-manager/src/lib/shared/filter/filter.component.scss
@@ -1,5 +1,5 @@
 .filter-container {
-  width: 14.71vw;
+  width: 14.9vw;
   height: 605px;
   background-color: #fafafa;
   border: solid 1px rgb(194, 193, 193);
