+++ src/app/models/availability/availability.spec.ts
@@ -1,0 +1,7 @@
+import { Availability } from './availability';
+
+describe('Availability', () => {
+  it('should create an instance', () => {
+    expect(new Availability()).toBeTruthy();
+  });
+});
