+++ src/app/models/campaigns/campaign.spec.ts
@@ -1,0 +1,7 @@
+import { Campaign } from './campaign';
+
+describe('Campaign', () => {
+  it('should create an instance', () => {
+    expect(new Campaign()).toBeTruthy();
+  });
+});
