+++ src/app/models/campaigns/campaign.ts
