+++ src/app/models/enums/available-status.ts
@@ -1,0 +1,4 @@
+export enum AvailableStatus {
+  AVAILABLE = 'AVAILABLE',
+  NOT_AVAILABLE = 'NOT_AVAILABLE',
+}
