+++ src/app/models/enums/priority.ts
@@ -1,0 +1,5 @@
+export enum Priority {
+  LOW = 'LOW',
+  MEDIUM = 'MEDIUM',
+  HIGH = 'HIGH',
+}
