+++ src/app/models/enums/status.ts
@@ -1,0 +1,6 @@
+export enum Status {
+  DEFINED = 'DEFINED',
+  IN_PROGRESS = 'IN_PROGRESS',
+  COMPLETED = 'COMPLETED',
+  ON_HOLD = 'ON_HOLD',
+}
