+++ src/app/models/leave/leave.spec.ts
@@ -1,0 +1,7 @@
+import { Leave } from './leave';
+
+describe('Leave', () => {
+  it('should create an instance', () => {
+    expect(new Leave()).toBeTruthy();
+  });
+});
