+++ src/app/models/leave/leave.ts
@@ -1,0 +1,7 @@
+export class Leave {
+  public constructor(
+    public leaveId: number,
+    public startDate: Date,
+    public endDate: Date
+  ) {}
+}
