+++ src/app/models/projects/project.spec.ts
@@ -1,0 +1,7 @@
+import { Project } from './project';
+
+describe('Project', () => {
+  it('should create an instance', () => {
+    expect(new Project()).toBeTruthy();
+  });
+});
