+++ src/app/models/resources/resource.spec.ts
@@ -1,0 +1,7 @@
+import { Resource } from './resource';
+
+describe('Resource', () => {
+  it('should create an instance', () => {
+    expect(new Resource()).toBeTruthy();
+  });
+});
