+++ src/app/models/resources/resource.ts
