+++ src/app/models/tasks/task.spec.ts
@@ -1,0 +1,7 @@
+import { Task } from './task';
+
+describe('Task', () => {
+  it('should create an instance', () => {
+    expect(new Task()).toBeTruthy();
+  });
+});
